+++ components/badge/ScrollNumber.jsx
@@ -36,10 +36,7 @@ class AntScrollNumber extends React.Component {
   }
 
   componentWillReceiveProps(nextProps) {
-    if ('count' in nextProps && nextProps.count) {
-      if (this.lastCount === this.state.count) {
-        return;
-      }
+    if ('count' in nextProps) {
       this.lastCount = this.state.count;
       // 复原数字初始位置
       this.setState({